Rename renderer module to display
