+++ client/src/display/graph.rs
