+++ client/src/display/model.rs
