+++ client/src/display/renderer.rs
