+++ client/src/display/scene.rs
